* SYNC [parser && scanner] sync hocon fixed.

+++ include/nng/supplemental/nanolib/parser.h
@@ -81,7 +81,7 @@ union YYSTYPE
 {
 #line 25 "parser.y"
 
-    int intval;
+    double intval;
     double floatval;
     char *strval;
     struct cJSON *jsonval;
